Test: 인벤토리 출력 테스트

+++ Source/ARPG_Hunter/Core/Subsystem/PlayerManager.cpp
@@ -48,7 +48,7 @@ void UPlayerManager::AddGold(uint32 _amount)
 	Gold.OnValueChanged.Broadcast(Gold.Value);
 }
 
-void UPlayerManager::AddItem(const FName& _itemID, uint16 _amount)
+void UPlayerManager::AddItem(const FName& _itemID, int32 _amount)
 {
 	//아이템 추가
 	uint8 Index;
